requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.2.2
crashtest==0.4.1
distlib==0.3.9
dulwich==0.21.7
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
narwhals==1.34.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pexpect==4.9.0
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==5.29.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
RapidFuzz==3.12.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
trove-classifiers==2025.3.19.19
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.35.0
virtualenv==20.29.3
xattr==1.1.4
zipp==3.21.0
